Add rechargeLevel and interestThreshold parameters

+++ Time Budget/AppState.swift
@@ -12,8 +12,8 @@ import Combine
 let longPressDuration = 0.15
 let longPressMaxDrift: CGFloat = 0.1
 let listViewExtension: CGFloat = 200
-let interestThreshold: Int = -1
-let rechargeLevel: Int = 2
+let interestThreshold: Float = -1
+let rechargeLevel: Float = 2
 
 struct AppState {
 	var currentPosition: Int? = nil
